Versão beta'

+++ src/main/java/com/iuri/pol/avaliacaotecnica/repository/ProcessoRepository.java
@@ -7,7 +7,7 @@ import com.iuri.pol.avaliacaotecnica.modelo.Processo;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProcessoRepository extends JpaRepository<Processo, Integer> {
+public interface ProcessoRepository extends JpaRepository<Processo, Long> {
 
     public Optional<List<Processo>> findAllByNmReu(String nmReu);
